# services/engine/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chatbot Engine API
  description: |
    REST API for the Chatbot Engine — an NLP-powered chatbot platform that supports
    multi-group configuration, intent recognition, query execution, learning feedback,
    and administrative management.
  version: 1.0.0
  contact:
    name: Chatbot Engine Team
servers:
  - url: http://localhost:4000
    description: Local development
  - url: '{productionUrl}'
    description: Production
    variables:
      productionUrl:
        default: https://chatbot-engine.example.com
        description: Production server URL
...
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        API key required for all `/api/admin/*` routes when the `ENGINE_API_KEY`
        environment variable is set. Can also be passed as a Bearer token in the
        Authorization header.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Internal server error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true

    # ── Chat ──────────────────────────────────────────────────────────────
...
    QueryStat:
      type: object
      properties:
        id:
          type: string
          example: stat_1710000000_ab12
        queryName:
          type: string
          example: open_tickets
        success:
          type: boolean
        durationMs:
          type: integer
        timestamp:
          type: string
          format: date-time
        filters:
          type: object
          additionalProperties:
            type: string
        error:
          type: string
    AggregatedStat:
      type: object
      properties:
        queryName:
          type: string
        totalExecutions:
          type: integer
        successCount:
          type: integer
        failureCount:
          type: integer
        failureRate:
          type: string
          example: '2.5%'
        avgDurationMs:
          type: integer
        p50Ms:
          type: integer
        p95Ms:
          type: integer

    # ── User Info ─────────────────────────────────────────────────────────
...
    ReviewItem:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        currentIntent:
          type: string
        confidence:
          type: number
        status:
          type: string
        timestamp:
          type: string
          format: date-time
    LearningStats:
      type: object
      properties:
        totalInteractions:
          type: integer
        autoLearnedCount:
          type: integer
        pendingReviewCount:
          type: integer

    # ── Admin: Settings ───────────────────────────────────────────────────
...
paths:
...
  /api/stats:
    get:
      tags: [Stats]
      summary: Get query execution statistics
      parameters:
        - name: queryName
          in: query
          schema:
            type: string
          description: Filter stats to a specific query
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Only include stats after this timestamp
      responses:
        '200':
          description: Statistics with aggregations
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueryStat'
                  aggregated:
                    type: array
                    items:
                      $ref: '#/components/schemas/AggregatedStat'
                  totalRecords:
                    type: integer
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Stats]
      summary: Record a query execution statistic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - queryName
              properties:
                queryName:
                  type: string
                  example: open_tickets
                success:
                  type: boolean
                  default: true
                durationMs:
                  type: integer
                  default: 0
                filters:
                  type: object
                  additionalProperties:
                    type: string
                error:
                  type: string
      responses:
        '201':
          description: Stat recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryStat'
        '400':
          description: Missing queryName
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: queryName is required
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # USER
  # ════════════════════════════════════════════════════════════════════════
...
  /api/admin/learning/review:
    get:
      tags: [Admin/Learning]
      summary: Get the learning review queue
      security:
        - ApiKeyAuth: []
      parameters:
        - name: groupId
          in: query
          schema:
            type: string
            default: default
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
          description: Filter by review status
      responses:
        '200':
          description: Review queue items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewItem'
                  total:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/learning/review/{id}/resolve:
    post:
      tags: [Admin/Learning]
      summary: Resolve a review item with the correct intent
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - correctIntent
              properties:
                correctIntent:
                  type: string
                  example: query.open_tickets
                groupId:
                  type: string
                  default: default
      responses:
        '200':
          description: Item resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Missing correctIntent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Review item not found or already resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/learning/review/{id}/dismiss:
    post:
      tags: [Admin/Learning]
      summary: Dismiss a review item
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                groupId:
                  type: string
                  default: default
      responses:
        '200':
          description: Item dismissed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Review item not found or already resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/learning/stats:
    get:
      tags: [Admin/Learning]
      summary: Get learning statistics
      security:
        - ApiKeyAuth: []
      parameters:
        - name: groupId
          in: query
          schema:
            type: string
            default: default
      responses:
        '200':
          description: Learning stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningStats'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/learning/auto-learned:
    get:
      tags: [Admin/Learning]
      summary: Get auto-learned items
      security:
        - ApiKeyAuth: []
      parameters:
        - name: groupId
          in: query
          schema:
            type: string
            default: default
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Auto-learned items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/learning/retrain:
    post:
      tags: [Admin/Learning]
      summary: Trigger NLP model retraining
      description: |
        Invalidates the cached NLP engine for the specified group, forcing a retrain
        on the next request. Use groupId "all" to retrain all groups.
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                groupId:
                  type: string
                  default: default
                  example: all
      responses:
        '200':
          description: Retrain triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  groupId:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/learning/process-signals:
    post:
      tags: [Admin/Learning]
      summary: Process pending feedback signals
      description: Processes accumulated user feedback signals and updates the learning model.
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                groupId:
                  type: string
                  default: default
      responses:
        '200':
          description: Signals processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: integer
                  newReviewItems:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ════════════════════════════════════════════════════════════════════════
  # ADMIN — SETTINGS
  # ════════════════════════════════════════════════════════════════════════
...